+++ Libplanet.Store/IStateStore.cs
@@ -16,9 +16,11 @@ namespace Libplanet.Store
         /// </summary>
         /// <param name="stateRootHash">The state root hash of the state root trie to get.
         /// If <see langword="null"/> is passed the empty state root trie is returned.</param>
+        /// <param name="readOnly">Whether the <see cref="ITrie"/> returned is read-only
+        /// in the sense that nothing gets written to the storage.</param>
         /// <returns>The state root trie of the <paramref name="stateRootHash"/>.
         /// If <see langword="null"/> is passed the empty state root trie is returned.</returns>
-        ITrie GetStateRoot(HashDigest<SHA256>? stateRootHash);
+        ITrie GetStateRoot(HashDigest<SHA256>? stateRootHash, bool readOnly = false);
 
         /// <summary>
         /// Prunes the states no more used from the state store.