+++ Libplanet.Store/TrieStateStore.cs
@@ -133,13 +133,17 @@ namespace Libplanet.Store
             _logger.Verbose("Finished {MethodName}()", nameof(CopyStates));
         }
 
-        /// <inheritdoc cref="IStateStore.GetStateRoot(HashDigest{SHA256}?)"/>
-        public ITrie GetStateRoot(HashDigest<SHA256>? stateRootHash) =>
-            new MerkleTrie(
-                StateKeyValueStore,
-                stateRootHash is { } hash ? new HashNode(hash) : null,
-                Secure
-            );
+        /// <inheritdoc cref="IStateStore.GetStateRoot(HashDigest{SHA256}?, bool)"/>
+        public ITrie GetStateRoot(HashDigest<SHA256>? stateRootHash, bool readOnly = false) =>
+            readOnly
+                ? new MerkleTrie(
+                    new VolatileKeyValueStore(StateKeyValueStore),
+                    stateRootHash is { } h1 ? new HashNode(h1) : null,
+                    Secure)
+                : new MerkleTrie(
+                    StateKeyValueStore,
+                    stateRootHash is { } h2 ? new HashNode(h2) : null,
+                    Secure);
 
         /// <inheritdoc cref="System.IDisposable.Dispose()"/>
         public void Dispose()